+++ js/app/src/routes/[...catchall]/+page.svelte
@@ -83,6 +83,7 @@
 	import { _ } from "svelte-i18n";
 	import { Client } from "@gradio/client";
 	import { page } from "$app/stores";
+	import { setupi18n } from "@gradio/core";
 
 	import { init } from "@huggingface/space-header";
 	import { browser } from "$app/environment";
@@ -253,11 +254,14 @@
 	let pending_deep_link_error = $state(false);
 
 	let gradio_dev_mode = "";
+	let i18n_ready = false;
+	setupi18n().then(() => {
+		i18n_ready = true;
+	});
 
 	onMount(async () => {
 		//@ts-ignore
 		config = data.config;
-		window.gradio_config = config;
 		window.gradio_config = data.config;
 		config = data.config;
 
@@ -265,7 +269,7 @@
 			pending_deep_link_error = true;
 		}
 
-		if (!app.config) {
+		if (!app?.config && !config.auth_required) {
 			throw new Error("Could not resolve app config");
 		}
 
@@ -442,6 +446,7 @@
 			root={config.root}
 			space_id={space}
 			{app_mode}
+			i18n={i18n_ready ? $_ : (s) => s}
 		/>
 	{:else if config && app}
 		<data.Render