+++ js/accordion/CHANGELOG.md
@@ -1,4 +1,11 @@
 # @gradio/accordion
+
+## 0.5.25
+
+### Dependency updates
+
+- @gradio/utils@0.10.4
+
 ## 0.5.25
 
 ### Features