+++ js/button/CHANGELOG.md
@@ -1,4 +1,11 @@
 # @gradio/button
+
+## 0.5.14
+
+### Dependency updates
+
+- @gradio/utils@0.10.4
+
 ## 0.5.14
 
 ### Features