+++ js/imageslider/CHANGELOG.md
@@ -1,4 +1,11 @@
 # @gradio/imageslider
+
+## 0.3.1
+
+### Dependency updates
+
+- @gradio/utils@0.10.4
+
 ## 0.3.1
 
 ### Features