+++ js/simpleimage/CHANGELOG.md
@@ -1,4 +1,11 @@
 # @gradio/simpleimage
+
+## 0.9.1
+
+### Dependency updates
+
+- @gradio/utils@0.10.4
+
 ## 0.9.1
 
 ### Features